Reorder the switching on the type of the arguments
Before we were checking for 'str' and '[]char' after all the integer types,
now we first check for 'str' and '[]char' and then for the integer types.
The idea is that printing/formatting 'str's is more common than that of
the smaller integer types (s16, u16, s8, u8), and also bool and char.